requirements.txt
django
django_filter
djangorestframework
mysqlclient